backend/gateway_service/requirements.txt
awscli==1.40.7
black==25.1.0
flake8==7.2.0
ipython==9.2.0
isort==6.0.1
mkdocs==1.6.1
python-dotenv==1.1.0
Django==5.1.5
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
drf-yasg==1.21.10
requests==2.32.3